Layout changes and recipe tweaks
Changed the layout of the ingredient table to two responsive columns (one for mobile).

Added hints 29 and 30.

Added functionality that takes into account the egg size to determine the water that needs to be added by the user.

Removed the default settings. Now only the settings added by the user will call hidden hints.